Matching Engine refactoring
- Break matching engine code into separate files
- Define interface for taskList
- Remove duplicate AddTask code for activity and decision tasks
